requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
gunicorn==23.0.0
gpiozero==2.0.1
pillow==11.1.0
scikit-image==0.25.1
numpy==1.24.2
cryptography==44.0.1
PyYAML==6.0.2
